chore: Moved definition/type/get-constructor-of.mts to utility/get-constructor-of.mts.
