+++ packages/architectura/src/definition/type/_index.mts
@@ -1,4 +1,4 @@
 export * from '../../core/server/definition/type/server-configuration.type.mjs';
 export * from '../../core/server/definition/type/server-instantiation.type.mjs';
+export * from '../../utility/get-constructor-of.mjs';
 export * from './constructor-of.type.mjs';
-export * from './get-constructor-of.mjs';
