+++ packages/architectura/src/utility/get-constructor-of.mts
@@ -1,4 +1,4 @@
-import type { ConstructorOf } from "./constructor-of.type.mjs";
+import type { ConstructorOf } from "../definition/type/constructor-of.type.mjs";
 
 function getConstructorOf<T extends object>(value: T): ConstructorOf<T>
 {
